+++ src/renderer/selectors.ts
@@ -2,7 +2,7 @@ import { createSelector, Selector } from 'reselect';
 import * as Cesium from 'cesium';
 
 import { requireElectron } from './electron';
-import { DEFAULT_BASE_MAP, DEFAULT_BASE_MAP_ID } from './state-util';
+import { canOpenDataSource, DEFAULT_BASE_MAP, DEFAULT_BASE_MAP_ID } from './state-util';
 import {
     BaseMapState,
     ColorMapCategoryState,
@@ -359,6 +359,7 @@ export const selectedDataSourceIdSelector = (state: State) => state.session.sele
 export const dataSourceFilterExprSelector = (state: State) => state.session.dataSourceFilterExpr;
 export const dataSourceListHeightSelector = (state: State) => state.session.dataSourceListHeight;
 export const showDataSourceDetailsSelector = (state: State) => state.session.showDataSourceDetails;
+export const showAllDataSourcesSelector = (state: State): boolean => state.session.showAllDataSources;
 export const showDataSourceIDsSelector = (state: State): boolean => state.session.showDataSourceIDs;
 export const showDataStoreDescriptionSelector = (state: State): boolean => state.session.showDataStoreDescription;
 export const showDataStoreNoticesSelector = (state: State): boolean => state.session.showDataStoreNotices;
@@ -390,35 +391,34 @@ export const filteredDataSourcesSelector = createSelector<State, DataSourceState
     boolean>(
     selectedDataSourcesSelector,
     dataSourceFilterExprSelector,
-    showDataSourceIDsSelector,
-    (selectedDataSources, dataSourceFilterExpr, showDataSourceTitles) => {
+    showAllDataSourcesSelector,
+    (selectedDataSources, dataSourceFilterExpr, showAllDataSources) => {
         const hasDataSources = selectedDataSources && selectedDataSources.length;
         const hasFilterExpr = dataSourceFilterExpr && dataSourceFilterExpr !== '';
-        if (hasDataSources && hasFilterExpr) {
+        if (hasDataSources && (hasFilterExpr || !showAllDataSources)) {
             const dataSourceFilterExprLC = dataSourceFilterExpr.toLowerCase();
             const parts = dataSourceFilterExprLC.split(' ');
-            const dsMatcher = showDataSourceTitles ? matchesIdOrTitle : matchesId;
-            return selectedDataSources.filter(ds => dsMatcher(ds, parts));
+            return selectedDataSources.filter(ds => matchesIdOrTitle(ds, showAllDataSources, parts));
         }
         return selectedDataSources;
     }
 );
 
-function matchesId(ds: DataSourceState, parts: string[]) {
+function matchesIdOrTitle(ds: DataSourceState, showAllDataSources: boolean, searchWords: string[]) {
+    let isCandidate = showAllDataSources || canOpenDataSource(ds);
+    if (!isCandidate) {
+        return false;
+    }
     const id = ds.id.toLowerCase();
-    return parts.every(part => id.includes(part));
-}
-
-function matchesIdOrTitle(ds: DataSourceState, parts: string[]) {
-    if (matchesId(ds, parts)) {
+    if (searchWords.every(word => id.includes(word))) {
         return true;
     }
-    let title = ds.title || (ds.meta_info ? ds.meta_info.title : null);
+    let title = ds.title || (ds.metaInfo ? ds.metaInfo.title : null);
     if (!title || title === '') {
         return false;
     }
     title = title.toLowerCase();
-    return parts.every(part => title.includes(part));
+    return searchWords.every(word => title.includes(word));
 }
 
 export const selectedDataSourceSelector = createSelector<State, DataSourceState | null, DataSourceState[] | null,