+++ src/renderer/state-util.ts
@@ -2,7 +2,7 @@ import * as Cesium from 'cesium';
 import { IconName } from '@blueprintjs/core';
 
 import {
-    AnimationViewDataState, BaseMapState, DataStoreState,
+    AnimationViewDataState, BaseMapState, DataSourceState, DataSourceVerificationFlags, DataStoreState,
     DimSizes,
     FigureViewDataState,
     LayerState,
@@ -81,6 +81,53 @@ export function isLocalDataStore(dataStore: DataStoreState | null) {
 export function isRemoteDataStore(dataStore: DataStoreState | null) {
     return dataStore && dataStore.id !== 'local' && !dataStore.isLocal;
 }
+
+export interface DataSourceUrls {
+    catalogUrl?: string;
+    infoUrl?: string;
+}
+
+export function getDataSourceUrls(dataSource: DataSourceState): DataSourceUrls {
+    const metaInfo = dataSource.metaInfo;
+    let catalogUrl;
+    let infoUrl;
+    if (metaInfo) {
+        catalogUrl = metaInfo.catalog_url || metaInfo.catalogue_url;
+        if (!catalogUrl && dataSource.id.includes("esacci") && metaInfo.uuid) {
+            catalogUrl = `https://catalogue.ceda.ac.uk/uuid/${metaInfo.uuid}`;
+        }
+        infoUrl = metaInfo.info_url || metaInfo.url;
+    }
+    return {catalogUrl, infoUrl};
+}
+
+export function canOpenDataSource(dataSource: DataSourceState) {
+    return _checkDataSource(dataSource, 'open', 'map', 'cache');
+}
+
+export function canCacheDataSource(dataSource: DataSourceState) {
+    return _checkDataSource(dataSource, 'cache');
+}
+
+function _checkDataSource(dataSource: DataSourceState, ...flags: DataSourceVerificationFlags[]): boolean {
+    if (Array.isArray(dataSource.verificationFlags)) {
+        // dataSource.verificationFlags have been introduced for ESA CCI datasets only
+        const s = new Set<string>(dataSource.verificationFlags);
+        for (let flag of flags) {
+            if (s.has(flag)) {
+                return true;
+            }
+        }
+        return false;
+    }
+    if (isString(dataSource.typeSpecifier)) {
+        // dataSource.typeSpecifier have been introduced for ESA CCI datasets only
+        return dataSource.typeSpecifier.startsWith("dataset");
+    }
+    // We assume, we can open all other (non ESA CCI) datasets
+    return true;
+}
+
 
 export function getTileUrl(baseUrl: string, baseDir: string, layer: VariableImageLayerState): string {
     return baseUrl + `ws/res/tile/${encodeURIComponent(baseDir)}/${layer.resId}/{z}/{y}/{x}.png?`